+++ glean/tests/core/metrics/timespan.spec.ts
@@ -4,6 +4,7 @@
 
 import assert from "assert";
 import sinon from "sinon";
+import type { SinonStub } from "sinon";
 
 import { Context } from "../../../src/core/context";
 import Glean from "../../../src/core/glean";
@@ -15,9 +16,11 @@ const sandbox = sinon.createSandbox();
 
 describe("TimespanMetric", function() {
   const testAppId = `gleanjs.test.${this.title}`;
+  let fakeNow: SinonStub;
 
   beforeEach(async function() {
     await Glean.testResetGlean(testAppId);
+    fakeNow = typeof performance === "undefined" ? sandbox.stub(Date, "now") : sandbox.stub(performance, "now");
   });
 
   afterEach(function () {
@@ -76,9 +79,8 @@ describe("TimespanMetric", function() {
   });
 
   it("ping payload is correct", async function() {
-    const fakeDateNow = sandbox.stub(Date, "now");
-    fakeDateNow.onCall(0).callsFake(() => 0);
-    fakeDateNow.onCall(1).callsFake(() => 100);
+    fakeNow.onCall(0).callsFake(() => 0);
+    fakeNow.onCall(1).callsFake(() => 100);
 
     const metric = new TimespanMetricType({
       category: "aCategory",
@@ -101,9 +103,8 @@ describe("TimespanMetric", function() {
   });
 
   it("recording APIs properly sets the value in all pings", async function() {
-    const fakeDateNow = sandbox.stub(Date, "now");
-    fakeDateNow.onCall(0).callsFake(() => 0);
-    fakeDateNow.onCall(1).callsFake(() => 100);
+    fakeNow.onCall(0).callsFake(() => 0);
+    fakeNow.onCall(1).callsFake(() => 100);
 
     const metric = new TimespanMetricType({
       category: "aCategory",
@@ -152,35 +153,34 @@ describe("TimespanMetric", function() {
       },
     ];
 
-    for (const testCase of testCases) {
-      const fakeDateNow = sandbox.stub(Date, "now");
-      fakeDateNow.onCall(0).callsFake(() => 0);
-      fakeDateNow.onCall(1).callsFake(() => 3600000); // One hour.
+    for (let i = 0; i < testCases.length; i++) {
+      fakeNow.onCall(0).callsFake(() => 0);
+      fakeNow.onCall(1).callsFake(() => 3600000); // One hour.
 
       const metric = new TimespanMetricType({
         category: "aCategory",
-        name: `aDatetimeMetric_${testCase.unit}`,
+        name: `aDatetimeMetric_${testCases[i].unit}`,
         sendInPings: ["aPing"],
         lifetime: Lifetime.Ping,
         disabled: false
-      }, testCase.unit);
+      }, testCases[i].unit);
 
       metric.start();
       metric.stop();
-      assert.strictEqual(await metric.testGetValue("aPing"), testCase.expected);
+      assert.strictEqual(await metric.testGetValue("aPing"), testCases[i].expected);
 
       sandbox.restore();
+      fakeNow = performance ? sandbox.stub(performance, "now") : sandbox.stub(Date, "now");
     }
   });
 
   it("second timer run is skipped", async function() {
-    const fakeDateNow = sandbox.stub(Date, "now");
     // First check, duration: 100
-    fakeDateNow.onCall(0).callsFake(() => 0);
-    fakeDateNow.onCall(1).callsFake(() => 100);
+    fakeNow.onCall(0).callsFake(() => 0);
+    fakeNow.onCall(1).callsFake(() => 100);
     // Second check, duration 99
-    fakeDateNow.onCall(2).callsFake(() => 101);
-    fakeDateNow.onCall(3).callsFake(() => 200);
+    fakeNow.onCall(2).callsFake(() => 101);
+    fakeNow.onCall(3).callsFake(() => 200);
 
     // TODO: check number of recorded errors instead once Bug 1682574 is resolved.
     const consoleErrorSpy = sandbox.spy(console, "error");
@@ -226,9 +226,8 @@ describe("TimespanMetric", function() {
   });
 
   it("nothing is stored before stop", async function() {
-    const fakeDateNow = sandbox.stub(Date, "now");
-    fakeDateNow.onCall(0).callsFake(() => 0);
-    fakeDateNow.onCall(1).callsFake(() => 100);
+    fakeNow.onCall(0).callsFake(() => 0);
+    fakeNow.onCall(1).callsFake(() => 100);
 
     const metric = new TimespanMetricType({
       category: "aCategory",
@@ -274,13 +273,12 @@ describe("TimespanMetric", function() {
     // Nothing should have been recorded.
     assert.strictEqual(await metric.testGetValue("aPing"), undefined);
 
-    // TODO: Make sure also incalid state error was recorded.
+    // TODO: Make sure also invalid state error was recorded.
   });
 
   it("time cannot go backwards", async function() {
-    const fakeDateNow = sandbox.stub(Date, "now");
-    fakeDateNow.onCall(0).callsFake(() => 100);
-    fakeDateNow.onCall(1).callsFake(() => 0);
+    fakeNow.onCall(0).callsFake(() => 100);
+    fakeNow.onCall(1).callsFake(() => 0);
 
     const metric = new TimespanMetricType({
       category: "aCategory",
@@ -298,5 +296,40 @@ describe("TimespanMetric", function() {
     assert.strictEqual(await metric.testGetValue("aPing"), undefined);
 
     assert.deepStrictEqual(consoleErrorSpy.callCount, 1);
+  });
+
+  it("setting raw time works correctly", async function () {
+    const metric = new TimespanMetricType({
+      category: "aCategory",
+      name: "aTimespan",
+      sendInPings: ["aPing"],
+      lifetime: Lifetime.Ping,
+      disabled: false
+    }, "millisecond");
+
+    metric.setRawNanos(1000000); // 1ms
+    assert.strictEqual(await metric.testGetValue("aPing"), 1);
+  });
+
+  it("setting raw time does nothing when times is running", async function () {
+    fakeNow.onCall(0).callsFake(() => 0);
+    fakeNow.onCall(1).callsFake(() => 100);
+
+    const metric = new TimespanMetricType({
+      category: "aCategory",
+      name: "aTimespan",
+      sendInPings: ["aPing"],
+      lifetime: Lifetime.Ping,
+      disabled: false
+    }, "millisecond");
+
+    metric.start();
+    metric.setRawNanos(42);
+    metric.stop();
+
+    // We expect the start/stop value, not the raw value.
+    assert.strictEqual(await metric.testGetValue("aPing"), 100);
+
+    // TODO: check number of recorded errors instead once Bug 1682574 is resolved.
   });
 });